Fold GeneratingMetadataManager into MetadataManager
Long time ago, `MetadataManager` had two descendants: the `GeneratingMetadataManager` and another one that I forgot the name of. It was used for .NET Native and didn't generate metadata, because ILTransforms did.

Fold the classes together.
