+++ src/coreclr/tools/aot/ILCompiler.Compiler/Compiler/MetadataManager.cs
@@ -3,11 +3,14 @@
 
 using System;
 using System.Collections.Generic;
+using System.IO;
+using System.Text;
 
 using Internal.TypeSystem;
 
 using ILCompiler.DependencyAnalysis;
 using ILCompiler.DependencyAnalysisFramework;
+using ILCompiler.Metadata;
 
 using Debug = System.Diagnostics.Debug;
 using ReadyToRunSectionType = Internal.Runtime.ReadyToRunSectionType;
@@ -17,9 +20,12 @@ using CombinedDependencyList = System.Collections.Generic.List<ILCompiler.Depend
 using CombinedDependencyListEntry = ILCompiler.DependencyAnalysisFramework.DependencyNodeCore<ILCompiler.DependencyAnalysis.NodeFactory>.CombinedDependencyListEntry;
 using MethodIL = Internal.IL.MethodIL;
 using CustomAttributeValue = System.Reflection.Metadata.CustomAttributeValue<Internal.TypeSystem.TypeDesc>;
+using MethodSignature = Internal.TypeSystem.MethodSignature;
 
 using MetadataRecord = Internal.Metadata.NativeFormat.Writer.MetadataRecord;
+using MetadataWriter = Internal.Metadata.NativeFormat.Writer.MetadataWriter;
 using TypeReference = Internal.Metadata.NativeFormat.Writer.TypeReference;
+using Field = Internal.Metadata.NativeFormat.Writer.Field;
 using TypeSpecification = Internal.Metadata.NativeFormat.Writer.TypeSpecification;
 using ConstantStringValue = Internal.Metadata.NativeFormat.Writer.ConstantStringValue;
 using TypeInstantiationSignature = Internal.Metadata.NativeFormat.Writer.TypeInstantiationSignature;
@@ -43,6 +49,8 @@ namespace ILCompiler
         private List<MetadataMapping<FieldDesc>> _fieldMappings;
         private List<MetadataMapping<MethodDesc>> _methodMappings;
         private List<StackTraceMapping> _stackTraceMappings;
+        protected readonly string _metadataLogFile;
+        protected readonly StackTraceEmissionPolicy _stackTraceEmissionPolicy;
 
         protected readonly CompilerTypeSystemContext _typeSystemContext;
         protected readonly MetadataBlockingPolicy _blockingPolicy;
@@ -75,7 +83,8 @@ namespace ILCompiler
         internal NativeLayoutInfoNode NativeLayoutInfo { get; private set; }
 
         public MetadataManager(CompilerTypeSystemContext typeSystemContext, MetadataBlockingPolicy blockingPolicy,
-            ManifestResourceBlockingPolicy resourceBlockingPolicy, DynamicInvokeThunkGenerationPolicy dynamicInvokeThunkGenerationPolicy,
+            ManifestResourceBlockingPolicy resourceBlockingPolicy, string logFile, StackTraceEmissionPolicy stackTracePolicy,
+            DynamicInvokeThunkGenerationPolicy dynamicInvokeThunkGenerationPolicy,
             MetadataManagerOptions options)
         {
             _typeSystemContext = typeSystemContext;
@@ -83,6 +92,8 @@ namespace ILCompiler
             _resourceBlockingPolicy = resourceBlockingPolicy;
             _dynamicInvokeThunkGenerationPolicy = dynamicInvokeThunkGenerationPolicy;
             _options = options;
+            _metadataLogFile = logFile;
+            _stackTraceEmissionPolicy = stackTracePolicy;
         }
 
         public bool IsDataDehydrated => (_options & MetadataManagerOptions.DehydrateData) != 0;
@@ -562,6 +573,12 @@ namespace ILCompiler
         }
 
         /// <summary>
+        /// Gets a list of fields that got "compiled" and are eligible for a runtime mapping.
+        /// </summary>
+        /// <returns></returns>
+        protected abstract IEnumerable<FieldDesc> GetFieldsWithRuntimeMapping();
+
+        /// <summary>
         /// This method is an extension point that can provide additional metadata-based dependencies to generated method bodies.
         /// </summary>
         public void GetDependenciesDueToMethodCodePresence(ref DependencyList dependencies, NodeFactory factory, MethodDesc method, MethodIL methodIL)
@@ -601,18 +618,28 @@ namespace ILCompiler
         /// Given that a method is invokable, if it is inserted into the reflection invoke table
         /// will it use a method token to be referenced, or not?
         /// </summary>
-        public abstract bool WillUseMetadataTokenToReferenceMethod(MethodDesc method);
+        public bool WillUseMetadataTokenToReferenceMethod(MethodDesc method)
+        {
+            return (GetMetadataCategory(method) & MetadataCategory.Description) != 0;
+        }
 
         /// <summary>
         /// Given that a method is invokable, if it is inserted into the reflection invoke table
         /// will it use a field token to be referenced, or not?
         /// </summary>
-        public abstract bool WillUseMetadataTokenToReferenceField(FieldDesc field);
+        public bool WillUseMetadataTokenToReferenceField(FieldDesc field)
+        {
+            return (GetMetadataCategory(field) & MetadataCategory.Description) != 0;
+        }
 
         /// <summary>
         /// Gets a stub that can be used to reflection-invoke a method with a given signature.
         /// </summary>
-        public abstract MethodDesc GetReflectionInvokeStub(MethodDesc method);
+        public MethodDesc GetReflectionInvokeStub(MethodDesc method)
+        {
+            return _typeSystemContext.GetDynamicInvokeThunk(method.Signature,
+                !method.Signature.IsStatic && method.OwningType.IsValueType);
+        }
 
         protected void EnsureMetadataGenerated(NodeFactory factory)
         {
@@ -634,6 +661,168 @@ namespace ILCompiler
                                                 out List<MetadataMapping<MethodDesc>> methodMappings,
                                                 out List<MetadataMapping<FieldDesc>> fieldMappings,
                                                 out List<StackTraceMapping> stackTraceMapping);
+
+        protected void ComputeMetadata<TPolicy>(
+            TPolicy policy,
+            NodeFactory factory,
+            out byte[] metadataBlob,
+            out List<MetadataMapping<MetadataType>> typeMappings,
+            out List<MetadataMapping<MethodDesc>> methodMappings,
+            out List<MetadataMapping<FieldDesc>> fieldMappings,
+            out List<StackTraceMapping> stackTraceMapping) where TPolicy : struct, IMetadataPolicy
+        {
+            var transformed = MetadataTransform.Run(policy, GetCompilationModulesWithMetadata());
+            MetadataTransform transform = transformed.Transform;
+
+            // Generate metadata blob
+            var writer = new MetadataWriter();
+            writer.ScopeDefinitions.AddRange(transformed.Scopes);
+
+            // Generate entries in the blob for methods that will be necessary for stack trace purposes.
+            var stackTraceRecords = new List<StackTraceRecordData>();
+            foreach (var methodBody in GetCompiledMethodBodies())
+            {
+                MethodDesc method = methodBody.Method;
+
+                MethodDesc typicalMethod = method.GetTypicalMethodDefinition();
+
+                // Methods that will end up in the reflection invoke table should not have an entry in stack trace table
+                // We'll try looking them up in reflection data at runtime.
+                if (transformed.GetTransformedMethodDefinition(typicalMethod) != null &&
+                    ShouldMethodBeInInvokeMap(method) &&
+                    (GetMetadataCategory(method) & MetadataCategory.RuntimeMapping) != 0)
+                    continue;
+
+                MethodStackTraceVisibilityFlags stackVisibility = _stackTraceEmissionPolicy.GetMethodVisibility(method);
+                bool isHidden = (stackVisibility & MethodStackTraceVisibilityFlags.IsHidden) != 0;
+
+                if ((stackVisibility & MethodStackTraceVisibilityFlags.HasMetadata) != 0)
+                {
+                    StackTraceRecordData record = CreateStackTraceRecord(transform, method, isHidden);
+
+                    stackTraceRecords.Add(record);
+
+                    writer.AdditionalRootRecords.Add(record.OwningType);
+                    writer.AdditionalRootRecords.Add(record.MethodName);
+                    writer.AdditionalRootRecords.Add(record.MethodSignature);
+                    writer.AdditionalRootRecords.Add(record.MethodInstantiationArgumentCollection);
+                }
+                else if (isHidden)
+                {
+                    stackTraceRecords.Add(new StackTraceRecordData(method, null, null, null, null, isHidden));
+                }
+            }
+
+            var ms = new MemoryStream();
+
+            // .NET metadata is UTF-16 and UTF-16 contains code points that don't translate to UTF-8.
+            var noThrowUtf8Encoding = new UTF8Encoding(false, false);
+
+            using (var logWriter = _metadataLogFile != null ? new StreamWriter(File.Open(_metadataLogFile, FileMode.Create, FileAccess.Write, FileShare.Read), noThrowUtf8Encoding) : null)
+            {
+                writer.LogWriter = logWriter;
+                writer.Write(ms);
+            }
+
+            metadataBlob = ms.ToArray();
+
+            const int MaxAllowedMetadataOffset = 0xFFFFFF;
+            if (metadataBlob.Length > MaxAllowedMetadataOffset)
+            {
+                // Offset portion of metadata handles is limited to 16 MB.
+                throw new InvalidOperationException($"Metadata blob exceeded the addressing range (allowed: {MaxAllowedMetadataOffset}, actual: {metadataBlob.Length})");
+            }
+
+            typeMappings = new List<MetadataMapping<MetadataType>>();
+            methodMappings = new List<MetadataMapping<MethodDesc>>();
+            fieldMappings = new List<MetadataMapping<FieldDesc>>();
+            stackTraceMapping = new List<StackTraceMapping>();
+
+            // Generate type definition mappings
+            foreach (var type in factory.MetadataManager.GetTypesWithEETypes())
+            {
+                MetadataType definition = type.IsTypeDefinition ? type as MetadataType : null;
+                if (definition == null)
+                    continue;
+
+                MetadataRecord record = transformed.GetTransformedTypeDefinition(definition);
+
+                // Reflection requires that we maintain type identity. Even if we only generated a TypeReference record,
+                // if there is an MethodTable for it, we also need a mapping table entry for it.
+                record ??= transformed.GetTransformedTypeReference(definition);
+
+                if (record != null)
+                    typeMappings.Add(new MetadataMapping<MetadataType>(definition, writer.GetRecordHandle(record)));
+            }
+
+            foreach (var method in GetReflectableMethods())
+            {
+                if (method.IsGenericMethodDefinition || method.OwningType.IsGenericDefinition)
+                {
+                    // Generic definitions don't have runtime artifacts we would need to map to.
+                    continue;
+                }
+
+                if (method.GetCanonMethodTarget(CanonicalFormKind.Specific) != method)
+                {
+                    // Methods that are not in their canonical form are not interesting
+                    continue;
+                }
+
+                if (IsReflectionBlocked(method.Instantiation) || IsReflectionBlocked(method.OwningType.Instantiation))
+                    continue;
+
+                if ((GetMetadataCategory(method) & MetadataCategory.RuntimeMapping) == 0)
+                    continue;
+
+                MetadataRecord record = transformed.GetTransformedMethodDefinition(method.GetTypicalMethodDefinition());
+
+                if (record != null)
+                    methodMappings.Add(new MetadataMapping<MethodDesc>(method, writer.GetRecordHandle(record)));
+            }
+
+            HashSet<FieldDesc> canonicalFields = new HashSet<FieldDesc>();
+            foreach (var field in GetFieldsWithRuntimeMapping())
+            {
+                FieldDesc fieldToAdd = field;
+                TypeDesc canonOwningType = field.OwningType.ConvertToCanonForm(CanonicalFormKind.Specific);
+                if (canonOwningType.IsCanonicalSubtype(CanonicalFormKind.Any))
+                {
+                    FieldDesc canonField = _typeSystemContext.GetFieldForInstantiatedType(field.GetTypicalFieldDefinition(), (InstantiatedType)canonOwningType);
+
+                    // If we already added a canonically equivalent field, skip this one.
+                    if (!canonicalFields.Add(canonField))
+                        continue;
+
+                    fieldToAdd = canonField;
+                }
+
+                Field record = transformed.GetTransformedFieldDefinition(fieldToAdd.GetTypicalFieldDefinition());
+                if (record != null)
+                    fieldMappings.Add(new MetadataMapping<FieldDesc>(fieldToAdd, writer.GetRecordHandle(record)));
+            }
+
+            // Generate stack trace metadata mapping
+            foreach (var stackTraceRecord in stackTraceRecords)
+            {
+                if (stackTraceRecord.OwningType != null)
+                {
+                    StackTraceMapping mapping = new StackTraceMapping(
+                        stackTraceRecord.Method,
+                        writer.GetRecordHandle(stackTraceRecord.OwningType),
+                        writer.GetRecordHandle(stackTraceRecord.MethodSignature),
+                        writer.GetRecordHandle(stackTraceRecord.MethodName),
+                        stackTraceRecord.MethodInstantiationArgumentCollection != null ? writer.GetRecordHandle(stackTraceRecord.MethodInstantiationArgumentCollection) : 0,
+                        stackTraceRecord.IsHidden);
+                    stackTraceMapping.Add(mapping);
+                }
+                else
+                {
+                    Debug.Assert(stackTraceRecord.IsHidden);
+                    stackTraceMapping.Add(new StackTraceMapping(stackTraceRecord.Method, 0, 0, 0, 0, stackTraceRecord.IsHidden));
+                }
+            }
+        }
 
         protected StackTraceRecordData CreateStackTraceRecord(Metadata.MetadataTransform transform, MethodDesc method, bool isHidden)
         {
